+++ src/components/role/RoleBasic.vue
@@ -230,10 +230,27 @@
         this.getRoleById(row.id);
       },
       handleDelete(index, row) {
-        this.$store.dispatch('deleteRole', {'param': row.id})
-          .then(res => {
-            this.initRoleList();
-          })
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          this.$store.dispatch('deleteRole', {'param': row.id})
+            .then(res => {
+              if (res.code === 0) {
+                this.$message({
+                  type: 'success',
+                  message: '删除成功!'
+                });
+              }
+              this.initRoleList();
+            })
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除'
+          });
+        });
       },
       getRoleById(id) {
         this.$store.dispatch('getRoleById', {'param': id})
